Reuse code in Redis metricsets
