+++ metricbeat/module/redis/metricset.go
@@ -15,49 +15,52 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package keyspace
+package redis
 
 import (
-	"github.com/elastic/beats/libbeat/common"
-	"github.com/elastic/beats/libbeat/logp"
+	"time"
+
+	rd "github.com/garyburd/redigo/redis"
+
 	"github.com/elastic/beats/metricbeat/mb"
-	"github.com/elastic/beats/metricbeat/mb/parse"
-	"github.com/elastic/beats/metricbeat/module/redis"
 )
-
-var (
-	debugf = logp.MakeDebug("redis-keyspace")
-)
-
-func init() {
-	mb.Registry.MustAddMetricSet("redis", "keyspace", New,
-		mb.WithHostParser(parse.PassThruHostParser),
-		mb.DefaultMetricSet(),
-	)
-}
 
 // MetricSet for fetching Redis server information and statistics.
 type MetricSet struct {
-	*redis.MetricSet
+	mb.BaseMetricSet
+	pool *rd.Pool
 }
 
 // New creates new instance of MetricSet
-func New(base mb.BaseMetricSet) (mb.MetricSet, error) {
-	ms, err := redis.NewMetricSet(base)
+func NewMetricSet(base mb.BaseMetricSet) (*MetricSet, error) {
+	// Unpack additional configuration options.
+	config := struct {
+		IdleTimeout time.Duration `config:"idle_timeout"`
+		Network     string        `config:"network"`
+		MaxConn     int           `config:"maxconn" validate:"min=1"`
+		Password    string        `config:"password"`
+	}{
+		Network:  "tcp",
+		MaxConn:  10,
+		Password: "",
+	}
+	err := base.Module().UnpackConfig(&config)
 	if err != nil {
 		return nil, err
 	}
-	return &MetricSet{ms}, nil
+
+	return &MetricSet{
+		BaseMetricSet: base,
+		pool: CreatePool(base.Host(), config.Password, config.Network,
+			config.MaxConn, config.IdleTimeout, base.Module().Config().Timeout),
+	}, nil
 }
 
-// Fetch fetches metrics from Redis by issuing the INFO command.
-func (m *MetricSet) Fetch() ([]common.MapStr, error) {
-	// Fetch default INFO.
-	info, err := redis.FetchRedisInfo("keyspace", m.Connection())
-	if err != nil {
-		return nil, err
-	}
+func (m *MetricSet) Connection() rd.Conn {
+	return m.pool.Get()
+}
 
-	debugf("Redis INFO from %s: %+v", m.Host(), info)
-	return eventsMapping(info), nil
+// Close redis connections
+func (m *MetricSet) Close() error {
+	return m.pool.Close()
 }
